serving/requirements.txt
flask==3.1.1
prometheus-fastapi-instrumentator==7.1.0
prometheus-client==0.22.1
kafka-python==2.2.13
# for inference
scikit-learn==1.5.0
pandas==2.2.2
numpy==2.2.6
matplotlib==3.9.0
seaborn==0.13.2
mlflow==3.1.1
psutil==6.1.0
scipy==1.14.1